requirements.txt
# Core LangChain and LLM Integration
langchain==0.2.11
langchain-google-genai==1.0.5
langchain-qdrant==0.1.3
google-generativeai>=0.5.2,<0.6.0

# Vector Database
qdrant-client==1.11.0

# Web Search and Scraping
langchain-google-community>=0.0.9
beautifulsoup4==4.12.2
requests==2.31.0

# UI Framework
streamlit==1.28.1
streamlit-chat==0.1.1

# Data Processing
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0

# Document Processing
pypdf==3.17.1
python-docx==0.8.11
markdown==3.5.1

# Utilities
pydantic>=2.7.4,<3.0.0
typing-extensions==4.8.0
